feat(backend): fix tests error
