+++ backend/service/board.go
@@ -15,11 +15,11 @@ type Board struct {
 }
 
 type BoardService struct {
-	redisRepo repository.RedisRepositoryInterface
+	RedisRepo repository.RedisRepositoryInterface
 }
 func NewBoardService () *BoardService {
 	return &BoardService {
-		redisRepo: &repository.RedisRepository{},
+		RedisRepo: &repository.RedisRepository{},
 	}
 }
 
@@ -28,8 +28,8 @@ func (srv *BoardService) getRedisId(id string) string {
 }
 
 func (srv *BoardService) List() []Board {
-	ids := srv.redisRepo.Keys(srv.getRedisId("*"))
-	list := srv.redisRepo.JsonMget(ids)
+	ids := srv.RedisRepo.Keys(srv.getRedisId("*"))
+	list := srv.RedisRepo.JsonMget(ids)
 	boards := []Board{}
 	for _, v := range list {
 		board := Board{}
@@ -43,7 +43,7 @@ func (srv *BoardService) List() []Board {
 }
 
 func (srv *BoardService) Get(id string) Board {
-	data := srv.redisRepo.JsonGet(srv.getRedisId(id))
+	data := srv.RedisRepo.JsonGet(srv.getRedisId(id))
 	board := Board{}
 	err := json.Unmarshal(data, &board)
 	if err != nil {
@@ -56,17 +56,17 @@ func (srv *BoardService) Create(board Board) string {
 	uuidObj, _ := uuid.NewUUID()
 	id := uuidObj.String()
 	board.Id = id
-	srv.redisRepo.JsonSet(srv.getRedisId(id), board)
+	srv.RedisRepo.JsonSet(srv.getRedisId(id), board)
 	return id
 }
 
 func (srv *BoardService) Update(id string, board Board) string {
-	srv.redisRepo.JsonSet(srv.getRedisId(id), board)
+	srv.RedisRepo.JsonSet(srv.getRedisId(id), board)
 	return id
 }
 
 func (srv *BoardService) Delete(id string) {
-	srv.redisRepo.Delete(srv.getRedisId(id))
+	srv.RedisRepo.Delete(srv.getRedisId(id))
 }
 
 func (srv *BoardService) Archive(id string) {
